hsutil: custom widgets to create input ints and checkboxes from another than default values
